init commit

+++ gridproxy/model/model.v
@@ -2,7 +2,7 @@ module model
 
 import time { Time }
 
-type ByteUnit = i64
+type ByteUnit = u32
 
 pub fn (u ByteUnit) to_megabytes() f64 {
 	return f64(u) / 1e+6
@@ -16,7 +16,7 @@ pub fn (u ByteUnit) to_terabytes() f64 {
 	return f64(u) / 1e+12
 }
 
-type SecondUnit = i64
+type SecondUnit = u32
 
 pub fn (u SecondUnit) to_minutes() f64 {
 	return f64(u) / 60
@@ -30,7 +30,7 @@ pub fn (u SecondUnit) to_days() f64 {
 	return f64(u) / (60 * 60 * 24)
 }
 
-type UnixTime = i64
+type UnixTime = u32
 
 pub fn (t UnixTime) to_time() Time {
 	return time.unix(t)
